% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{allJobs}
\alias{allJobs}
\title{GMQL API web Service}
\usage{
allJobs(datasetName)
}
\arguments{
\item{username}{user name}

\item{password}{user password}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
