% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder-class.R
\name{BUILDER}
\alias{BUILDER}
\title{BUILDER object class}
\usage{
BUILDER()
}
\description{
BUILDER object available are:
\itemize{
\item{LEFT}
\item{RIGHT}
\item{CONTIG}
\item{INTERSECTION}
}
}
\details{
you never use a parent class BUILDER()
}
