% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{compileQuery.fromfile}
\alias{compileQuery.fromfile}
\title{GMQL API web Service}
\usage{
compileQuery.fromfile(url, filePath)
}
\arguments{
\item{url}{server address}

\item{filePath}{path of txt files where you wrote a query}
}
\description{
It compile a query from file
}
