% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{compileQuery}
\alias{compileQuery}
\title{Compile GMQL query}
\usage{
compileQuery(url, query)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{query}{text query}
}
\description{
It compiles a GMQL query
}
