% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{CONDITION}
\alias{CONDITION}
\title{CONDITION object class}
\usage{
CONDITION(value)
}
\description{
CONDITION object available are:
\itemize{
\item{DEF: DEFAULT}
\item{EXACT: EXACT}
\item{FULL: FULLNAME}
}
}
\details{
you never use a parent class CONDITION()
}
