% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{DESC}
\alias{DESC}
\title{ORDER object class}
\usage{
DESC(value)
}
\arguments{
\item{value}{single string identifying name of metadata attribute}
}
\value{
DESC ordering object
}
\description{
This class is used to create an ordering object for input to GMQL function
}
\details{
ORDER object available are:
\itemize{
\item{ASC}
\item{DESC}
}
}
