% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DGE}
\alias{DGE}
\title{DISTAL object class}
\usage{
DGE(value)
}
\value{
no returned value
}
\description{
DISTAL object available are:
\itemize{
\item{UP}
\item{DOWN}
\item{DGE}
\item{DLE}
\item{MD}
}
}
