% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distal-class.R
\name{DISTAL}
\alias{DISTAL}
\title{DISTAL object class}
\usage{
DISTAL(value)
}
\description{
DISTAL object available are:
\itemize{
\item{UP}
\item{DOWN}
\item{DGE}
\item{DLE}
\item{MD}
}
}
\details{
you never use a parent class DISTAL()
}
