% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download Dataset}
\usage{
downloadDataset(url, datasetName, path = getwd())
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}

\item{path}{local path folder where store dataset
by defualt is R working directory}
}
\description{
It donwloads private dataset as zip file from repository to local path specified
}
\details{
If error occured a specific error will be printed
}
\examples{

\dontrun{

login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
downloadDataset(url,<dataset_name>,path = <folder_path>)
}
}
