% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesToGMQL.R
\name{exportGMQL.gdm}
\alias{exportGMQL.gdm}
\title{Create GMQL dataset from GrangesList}
\usage{
exportGMQL.gdm(samples, dir_out)
}
\arguments{
\item{samples}{GrangesList}

\item{dir_out}{folder path where create a folder and write all the sample files}
}
\value{
no value return
}
\description{
It create Gmql dataset from GRangesList
All sample are in GDM (tab-separated values) file format
}
\details{
The GMQL dataset is made up by two differet file type

\itemize{
\item{metadata files: contains metadata associated to corrisponding sample}
\item{region files: contains many chromosome regions }
\item{region schema file: XML file contains region attribute (e.g. chr, start, end, pvalue)}
}
regions sample file and metadata file are associated through file name:
for example S_0001.gdm for regions file and S_0001.gdm.meta for its metadata
}
\examples{

library(GenomicRanges)
gr1 <- GRanges(seqnames = "chr2", ranges = IRanges(3, 6), strand = "+", score = 5L, GC = 0.45)
gr2 <- GRanges(seqnames = c("chr1", "chr1"),
ranges = IRanges(c(7,13), width = 3), strand = c("+", "-"), score = 3:4, GC = c(0.3, 0.5))
grl = GRangesList(gr1,gr2)
test_out_path <- system.file("example",package = "GMQL")
exportGMQL.gdm(grl,test_out_path)


}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gtf}}
}
