% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{extend}
\alias{extend}
\title{GMQL Operation: EXTEND}
\usage{
extend(metadata = NULL, input_data)
}
\arguments{
\item{metadata}{a list of element key = value, where value is c("aggregate function","value")}

\item{input_data}{string pointer returned from all GMQL function}
}
\description{
It generates new metadata attributes Am as result of aggregate
functions g applied to region attributes;
aggregate functions are applied sample by sample,
resulting tuples are triples with the sample identifier,
the attribute name Am, and the computed aggregate value.
}
\examples{
r = read(path)
e = extend(metadata = list(regionsCount = c("COUNT","")),input_data = r)
e = extend(metadata = list(sumValue = c("SUM","pvalue")),input_data = r)
e = extena(input_data = r)

}
