% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{flat}
\alias{flat}
\title{GMQL Operation: FLAT}
\usage{
flat(minAcc, maxAcc, groupBy = NULL, aggregates = NULL, input_data)
}
\description{
returns the contiguous region that starts from the first end and stops at
the last end of the regions which would contribute to each region of the COVER
}
