% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-class.R
\name{FULL}
\alias{FULL}
\title{CONDITION object class}
\usage{
FULL(value)
}
\value{
no returned value
}
\description{
CONDITION object available are:
\itemize{
\item{FULL: Fullname evaluation, two attributes match if they both end with value and,
if they have a further prefixes, the two prefix sequence are identical}
\item{DEF: Default evaluation, two attributes match if both end with value. }
\item{EXACT: Exact evaluation, only attributes exactly as value will match; no further prefixes are allowed. }
}
}
