% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{GMQLlogin}
\alias{GMQLlogin}
\title{GMQL API web Service}
\usage{
GMQLlogin(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{server address}

\item{username}{user name}

\item{password}{user password}
}
\value{
autentication Token
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
\examples{
GMQLlogin(url = http://...../...)
GMQLlogin(url, username="pippo",password="baudo")

}
