% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{GMQLlogout}
\alias{GMQLlogout}
\title{GMQL API web Service}
\usage{
GMQLlogout(url = "http://genomic.elet.polimi.it/gmql-rest/logout")
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
