% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{GRangesListFromGMQL.gdm}
\alias{GRangesListFromGMQL.gdm}
\title{create GrangesList from sample in GDM (delimited / tabulated) format file}
\usage{
GRangesListFromGMQL.gdm(datasetName = "DATA_SET_VAR_GDM/files")
}
\description{
create GrangesList from sample in GDM (delimited / tabulated) format file
}
