% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{GRangesListFromGMQL.gtf}
\alias{GRangesListFromGMQL.gtf}
\title{create GrangesList from sample in GTF format file}
\usage{
GRangesListFromGMQL.gtf(datasetName = "DATA_SET_VAR_GTF/files")
}
\description{
create GrangesList from sample in GTF format file
}
