% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesToGMQL.R
\name{GRangesToGMQL}
\alias{GRangesToGMQL}
\title{create GMQL dataset from Granges or GrangesList}
\usage{
GRangesToGMQL(..., dir_out, file_ouput_gtf = T)
}
\description{
create GMQL dataset from Granges or GrangesList
}
