% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{histogram}
\alias{histogram}
\title{GMQL Operation: HISTOGRAM}
\usage{
histogram(minAcc, maxAcc, groupBy = NULL, aggregates = NULL, input_data)
}
\description{
returns the non-overlapping regions contributing to the cover,
each with its accumulation index value, which is assigned to the AccIndex region attribute.
}
