% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gdm}
\alias{importGMQL.gdm}
\title{Create GrangesList from dataset GMQL in GDM (delimited / tabulated) format file}
\usage{
importGMQL.gdm(datasetName = "/Users/simone/Downloads/DATA_SET_VAR_GDM")
}
\arguments{
\item{datasetName}{dataset folder path}
}
\description{
Create GrangesList from dataset GMQL in GDM (delimited / tabulated) format file
}
\examples{
path <- "/..../..../..../DATA_SET_VAR_GDM"
grl <- importGMQL.gdm(path)

}
