% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gtf}
\alias{importGMQL.gtf}
\title{Create GrangesList from dataset GMQL in GTF format file}
\usage{
importGMQL.gtf(datasetName = "/Users/simone/Downloads/DATA_SET_VAR_GTF")
}
\arguments{
\item{datasetName}{dataset folder path}
}
\description{
Create GrangesList from dataset GMQL in GTF format file
}
\examples{

}
