% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{initGMQL}
\alias{initGMQL}
\title{Init GMQL Server}
\usage{
initGMQL(output_format)
}
\arguments{
\item{output_format}{single string identifies the output format of sample files.
Can be TAB, GTF, VCF or COLLECT
\itemize{
\item{TAB: tab delimited file format}
\item{GTF: file format used to hold information about gene structure.
It is a tab-delimited text format based on the general feature format}
\item{COLLECT: used for storing output in memory}
\item{VCF: Text file format (most likely stored in a compressed manner).
It contains meta-information lines, a header line, data lines each containing information about a position in the genome.}
}}
}
\description{
Initialize and run GMQL server for executing GMQL query
}
\details{
Instantiate in .GlobalEnv a scala singleton (Wrapper) used to call its scala methods
}
\examples{
\dontrun{

initGMQL("tab")
}

}
