% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join.R
\name{join}
\alias{join}
\title{GMQL Operation: JOIN}
\usage{
join(genometric_predicate = NULL, joinBy = NULL, output = CONTIG(),
  right_input_data, left_input_data)
}
\arguments{
\item{right_input_data}{"url-like" string taken from GMQL function}

\item{left_input_data}{"url-like" string taken from GMQL function}
}
\description{
It takes in input two datasets, respectively known as nchor (left) and experiment (right) and returns
a dataset of samples consisting of regions extracted from the operands according to the specified condition
(a.k.a genometric_predicate).
The number of generated output samples is the Cartesian product of the number of samples
in the anchor and in the experiment dataset (if joinBy is not specified).
The output metadata are the union of the input metadata, with their attribute names prefixed with
left or right respectively.
}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
