% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.R
\name{map}
\alias{map}
\title{GMQL Operation: MAP}
\usage{
map(left_input_data, right_input_data, aggregates = NULL, joinBy = NULL)
}
\arguments{
\item{left_input_data}{"url-like" string taken from GMQL function}

\item{right_input_data}{"url-like" string taken from GMQL function}
}
\description{
It computes, for each sample in the right dataset, aggregates over the values of the right regions
that intersect with a region in a left sample, for each region of each sample in the left dataset;
The number of generated output samples is the Cartesian product of the samples in the two input datasets;
each output sample has the same regions as the related input left sample, with their attributes and values,
plus the attributes computed as aggregates over right region values.
Output sample metadata are the union of the related input sample metadata,
whose attribute names are prefixed with "left" or "right" respectively.
}
\details{
When the joinby clause is present, only pairs of samples of left_input_data and of right_input_data with
metadata M1 and M2 respectively that satisfy the joinby condition are considered.

The clause consists of a list of metadata attribute names that must be present with equal values
in both M1 and  M2
}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
