% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{materialize}
\alias{materialize}
\title{GMQL Operation: MATERIALIZE}
\usage{
materialize(input_data, dir_out)
}
\arguments{
\item{input_data}{string pointer taken from GMQL function}

\item{dir_out}{out path folder for default is working directory}
}
\description{
It saved the content of a dataset,whose name can be specified,
that contains samples metadata and samples regions.
To preserve the content of any dataset generated during a GMQL query,
the dataset must be materialized.
Any dataset can be materialized, however the operation is time expensive;
for best performance, materialize the relevant data only.
}
\examples{

startGMQL()
r = read(path)
r2 = read(path2)
s = select(input_data = r)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = s)
materialize(input_data = m, dir_out = "/.../foldername")
materialize(s,"/.../foldername")

}
