% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{META_OPERATOR}
\alias{META_OPERATOR}
\title{META_OPERATOR object class}
\usage{
META_OPERATOR(value)
}
\description{
META_OPERATOR object available are:
\itemize{
\item{SUM}
\item{MIN}
\item{MAX}
\item{COUNT}
\item{BAG}
\item{AVG}
\item{STD}
\item{MEDIAN}
\item{Q1}
\item{Q2}
\item{Q3}
}
}
\details{
you never use a parent class META_OPERATOR()
}
