% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowse.R
\name{myQueries}
\alias{myQueries}
\title{GMQL API web Service}
\usage{
myQueries(url = "http://genomic.elet.polimi.it/gmql-rest/query")
}
\arguments{
\item{url}{server address}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
