% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{OPERATOR}
\alias{OPERATOR}
\title{OPERATOR object class}
\usage{
OPERATOR(value)
}
\description{
OPERATOR object available are:
\itemize{
\item{SUM}
\item{MIN}
\item{MAX}
\item{COUNT}
\item{BAG}
\item{AVG}
}
}
\details{
you never use a parent class OPERATOR()
}
