% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{order}
\alias{order}
\title{GMQL operation: ORDER}
\usage{
order(metadata_order = NULL, mtop = 0, mtopg = 0, regions_order = NULL,
  rtop = 0, rtopg = 0, input_data)
}
\arguments{
\item{metadata_order}{aaa}

\item{mtop}{0 is default meand that we not use it}

\item{mtopg}{0 is default meand that we not use it}

\item{regions_order}{aaa}

\item{rtop}{0 is default meand that we not use it}

\item{rtopg}{0 is default meand that we not use it}

\item{input_data}{aaaa}
}
\description{
It orders either samples, or regions, or both of them; order is ascending and or descending.
Sorted samples or regions have a new attribute Order, added to either metadata, or regions,
or both of them; the value of Order reflects the result of the sorting.
The input mtop = k extracts the first k samples or regions, the clause mtopg = k implicitly
considers the grouping by identical values of the first n − 1 ordering attributes
and then selects the first k samples or regions of each group
}
