% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-class.R
\name{PARSER}
\alias{PARSER}
\title{PARSER object class}
\usage{
PARSER()
}
\description{
PARSER object available are:
\itemize{
\item{BedParser}
\item{ANNParser}
\item{BroadProjParser}
\item{NarrowPeakParser}
\item{BasicParser}
\item{RnaSeqParser}
\item{CustomParser}
}
}
\details{
you never use a parent class PARSER()
}
