% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{project}
\alias{project}
\title{GMQL Operation: PROJECT}
\usage{
project(metadata = NULL, regions = NULL, input_data)
}
\arguments{
\item{input_data}{string pointer taken from GMQL function}

\item{predicate}{string made up by logical oepration: AND,OR,NOT}

\item{region}{region}

\item{semijoin}{semijoin}
}
\description{
It creates, from an existing dataset, a new dataset with all the samples from input dataset
but keeping for each sample in the input dataset only those metadata and/or region attributes
expressed in the operator parameter list.
Region coordinates and values of the remaining metadata remain equal to those in the input dataset.
It allows to:
a) Remove existing metadata and/or region attributes from a dataset;
b) Create new metadata and/or region attributes in the result
}
