% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{prova}
\alias{prova}
\title{Retrieve a Java instance of}
\usage{
prova()
}
\arguments{
\item{no}{param}
}
\value{
a Java instance
}
\description{
Retrieve a Java instance of SpellCorrector, with the training file
specified. Language model is trained before the instance is returned.
The spell corrector is adapted from Peter Norvig's demonstration.
}
