% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator-class.R
\name{Q2}
\alias{Q2}
\title{OPERATOR object class}
\usage{
Q2(value)
}
\description{
OPERATOR object available are:
\itemize{
\item{SUM}
\item{MIN}
\item{MAX}
\item{COUNT}
\item{BAG}
\item{AVG}
\item{STD}
\item{MEDIAN}
\item{Q1}
\item{Q2}
\item{Q3}
}
}
