% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(DatasetFolder, parser = "CustomParser", is_local = T)
}
\arguments{
\item{DatasetFolder}{folder path for GMQL dataset or datasetname on repository}

\item{parser}{string used to parsing dataset files
The Parser's available are:
\itemize{
\item{BedParser}
\item{ANNParser}
\item{BroadProjParser}
\item{BedParser}
\item{NarrowPeakParser}
\item{RnaSeqParser}
\item{CustomParser.}
}
Default is CustomParser.}

\item{local}{logical}
}
\value{
"url-like" string to dataset
}
\description{
Read a GMQL dataset or any other folder containig some homogenus sample
from disk, saving in Scala memory that can be referenced in R
}
\details{
Normally a GMQL dataset contains an XML schema file that contains
name of column header. (e.g chr, start, stop, strand)
The CustomParser read this XML schema; if you already know what kind of schema your files are use one of
the parser defined without reading any XML schema
}
\examples{

\dontrun{

initGMQL("gtf")
path = "/<path_to_your_folder>/<your_dataset_name>"
r = read(path)
r = read(path,"BedParser")
r = read(path,"RnaSeqParser")
}

}
