% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{register.GMQL}
\alias{register.GMQL}
\title{Sign up to GMQL}
\usage{
register.GMQL(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{name}{single string first name of the user (can contain space)}

\item{lastname}{single string last name of the user (can contain space)}

\item{mail}{single string email (as spacified in RFC format)}

\item{username}{single string name you want to authenticate with}

\item{password}{single string password you want to authenticate with}
}
\description{
Sign up to GMQL rest services suite
}
\details{
After registration you receive an authentication token.
As token remains vaild on server (until the next login / registration) a user can safely use a token for
a previous session as a convenience, this token is saved in Global environment to perform
subsequent REST call even on complete R restart (if is environemnt has been saved, of course ...)
If error occured a specific error will be printed
}
\examples{

\dontrun{

url <- <http_server_address>)
register.GMQL(url = url,"jonh","Doe","jonh@doe.com","JonhDoe46","password")
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
