% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{remote_processing}
\alias{remote_processing}
\title{Disable or Enable remote processing}
\usage{
remote_processing(is_remote)
}
\arguments{
\item{is_remote}{logical value, if TRUE you will set a remote query processing mode otherwise will be local}
}
\description{
It allows to enable or disable remote processing
}
\examples{
\dontrun{

}
""

}
