% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{runQuery.fromfile}
\alias{runQuery.fromfile}
\title{GMQL API web Service}
\usage{
runQuery.fromfile(url, fileName, filePath, output_gtf = T)
}
\arguments{
\item{url}{server address}

\item{fileName}{name of file}

\item{filePath}{path of txt files where you wrote a query}

\item{output_gtf}{logical: file format for sample generating from query
the possiblities are: GTF or TAB}
}
\description{
It runs a query from file
}
