% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{runQuery}
\alias{runQuery}
\title{Run GMQL query}
\usage{
runQuery(url, fileName, query, output_gtf = TRUE)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{fileName}{name of file}

\item{query}{text of query}

\item{output_gtf}{logical: file format for sample generating from query
the possiblities are: GTF or TAB}
}
\value{
no return value
}
\description{
It runs a GMQL query
}
\details{
If error occured a specific error will be printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
runQuery(PolimiUrl, "query_1", "DATA_SET_VAR = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATA_SET_VAR INTO RESULT_DS;", output_gtf = FALSE)

}
