% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{samplesFromDataset}
\alias{samplesFromDataset}
\title{GMQL API web Service}
\usage{
samplesFromDataset(datasetName)
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
