% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{saveQuery.fromfile}
\alias{saveQuery.fromfile}
\title{Save GMQL query from file}
\usage{
saveQuery.fromfile(url, queryName, filePath)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{queryName}{name of GMQL query}

\item{filePath}{local file path where you write GMQL query}
}
\description{
It saves the GMQL query into repository taken from file
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
saveQuery.fromfile(url,"prova1","<query_file_path>")
logout.GMQL()
}

}
\seealso{
\code{\link{saveQuery}}
}
