% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-class.R
\name{SCHEMA}
\alias{SCHEMA}
\title{SCHEMA object class}
\usage{
SCHEMA()
}
\description{
SCHEMA object available are:
\itemize{
\item{NARROWPEAK}
\item{BROADPEAK}
\item{VCF}
\item{BED}
\item{BEDGRAPH}
}
}
\details{
you never use a parent class SCHEMA()
}
