% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{showJobLog}
\alias{showJobLog}
\title{GMQL API web Service}
\usage{
showJobLog(url, job_id)
}
\arguments{
\item{url}{server address}

\item{job_id}{id of jobs}
}
\description{
It show a log for specific job
}
