% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showSamplesFromDataset}
\alias{showSamplesFromDataset}
\title{Show dataset samples}
\usage{
showSamplesFromDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}
}
\value{
list of samples in dataset.
Every sample in the list is identified by:
\itemize{
\item{id: id of sample}
\item{name: name of sample}
\item{path: sample repository path}
}
}
\description{
It show all sample from a specific GMQL dataset
}
\details{
If error occured a specific error will be printed
}
\examples{

\dontrun{
login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
list <- showSamplesFromDataset(url,<dataset_name>)
}
}
\seealso{
\code{\link{uploadSamples}}
}
