% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showSchemaFromDataset}
\alias{showSchemaFromDataset}
\title{GMQL API web Service}
\usage{
showSchemaFromDataset(datasetName)
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
