% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{summit}
\alias{summit}
\title{GMQL Operation: SUMMIT}
\usage{
summit(input_data, minAcc, maxAcc, groupBy = NULL, aggregates = NULL)
}
\arguments{
\item{input_data}{"url-like" string taken from GMQL function}

\item{minAcc}{minimum number of overlapping regions to be considered during execution
Normally must be > 0, we admit value 0 and -1 as special value:
\itemize{
\item {-1: means ALL and sets the minimum to the number of samples in the input dataset}
}}

\item{maxAcc}{maximum number of overlapping regions to be considered during execution
\itemize{
\item {0: means ANY and acts as a wildcard and can be used only as maxAcc value}
\item {-1: means ALL and sets the maximum to the number of samples in the input dataset}
}}

\item{groupBy}{a vector of strings specifying grouping criteria}

\item{aggregates}{a list of element in the form key = 'function_aggregate'.
'function_aggregate' is an object of class OPERATOR
The aggregate functions available are: MIN, MAX, SUM, BAG, AVG, COUNT.
Every operator accepts a string value, execet for COUNT that cannot have a value.
Argument of 'function_aggregate' must exist in schema
Two style are allowed}
}
\value{
"url-like" string
}
\description{
returns regions that start from a position
where the number of intersecting regions is not increasing afterwards and stops
at a position where either the number of intersecting regions decreases,
or it violates the max accumulation index).
}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
\seealso{
\code{\link{flat}} \code{\link{cover}} \code{\link{histogram}}
}
