% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{take}
\alias{take}
\title{GMQL Operation: TAKE}
\usage{
take(input_data, rows = 0L)
}
\arguments{
\item{input_data}{"url-like" string taken from GMQL function}

\item{rows}{number of rows that you want to retrieve an stored in memory
by default is 0 that means take all rows}
}
\value{
GrangesList with associated metadata
}
\description{
GMQL Operation: TAKE
}
\examples{

\dontrun{

library(rscala)

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)
s = select(input_data = r)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = s)
take(input_data = m, rows = 45)
}

}
