% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQL4TFarm.R
\name{TFARMemAtrix}
\alias{TFARMemAtrix}
\title{GrangesList to TFARM matrix}
\usage{
TFARMemAtrix(rangesList, metadata = NULL, metadata_prefix = NULL,
  regions = NULL)
}
\arguments{
\item{rangesList}{GrangesList object}

\item{metadata_prefix}{a a a a a a a}

\item{regions}{a a a a a a}
}
\value{
TFARM matrix
}
\description{
create matrix for TFARM package directly by GrangesList
}
\examples{


test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
TFARMemAtrix(test_path,regions = c("pvalue","peak"))


}
