% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{traceJob}
\alias{traceJob}
\title{Trace a job}
\usage{
traceJob(url, job_id)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{job_id}{single string id of the job}
}
\description{
It traces a specific current job
}
\details{
If error occured a specific error will be printed
}
