% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Union.R
\name{union}
\alias{union}
\title{GMQL Operation: UNION}
\usage{
union(left_input_data, right_input_data)
}
\arguments{
\item{left_input_data}{returned object from any GMQL function}

\item{right_input_data}{returned object from any GMQL function}
}
\value{
"url-like" string
}
\description{
It is used to integrate homogeneous or heterogeneous samples of two datasets within a single dataset
for each sample of either input dataset, a result sample is created as follows:
\itemize{
\item Metadata are the same as in the original sample.
\item Resulting schema is obtained by projecting the schema of the right dataset over the schema of the left one
(more properly, it will be performed by adding to the schema of the left dataset the region attributes of the right dataset
which are not identical to those of the left dataset)
\item Regions are the same (in coordinates and attribute values) as in the original sample.
Region attributes which are missing in an input dataset sample w.r.t. the merged schema are set to null.
}
For what concerns metadata, attributes of samples from the left (right) input dataset are prefixed
with the strings LEFT (RIGHT), so as to trace the dataset to which they originally belonged.
}
\examples{

\dontrun{

library(rscala)

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
test_path2 <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
r = read(test_path)
r2 = read(test_path2)
u = union(r2,r)
}
.

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
