% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{compileQuery.fromfile}
\alias{compileQuery.fromfile}
\title{Compile GMQL query from file}
\usage{
compileQuery.fromfile(url, filePath)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{filePath}{path of txt files where you wrote a query}
}
\value{
no return value
}
\description{
It compiles a GMQL query from file
}
\details{
If error occured a specific error will be printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
test_path <- system.file("example",package = "GMQL")
test_query <- file.path(test_path, "query1.txt")
compileQuery.fromfile(PolimiUrl,test_query)

}
