% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{compileQuery}
\alias{compileQuery}
\title{GMQL API web Service}
\usage{
compileQuery(url, query)
}
\arguments{
\item{url}{server address}

\item{query}{txt query}
}
\description{
It compile a query
}
