% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{compileQuery}
\alias{compileQuery}
\title{GMQL API web Service}
\usage{
compileQuery(url = "http://genomic.elet.polimi.it/gmql-rest/queries/compile",
  query)
}
\arguments{
\item{username}{user name}

\item{password}{user password}
}
\value{
authentication token
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
\details{
si predilige il testo della query
}
