% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{compileQuery}
\alias{compileQuery}
\title{Compile GMQL query}
\usage{
compileQuery(url, query)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{query}{text query}
}
\value{
no return value
}
\description{
It compiles a GMQL query
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
compileQuery(PolimiUrl, "DATA_SET_VAR = SELECT() HG19_TCGA_dnaseq;
MATERIALIZE DATA_SET_VAR INTO RESULT_DS;")

}
