% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{cover}
\alias{cover}
\title{GMQL Operation: COVER}
\usage{
cover(minAcc, maxAcc, groupBy = NULL, aggregates = NULL, input_data)
}
\description{
responds to the need of computing properties that reflect region’s intersections,
the operation with no grouping produces a single output sample,
and all the metadata attributes of the contributing input samples in result dataset
are assigned to the resulting single sample in input dataset
Regions of the result sample are built from the regions of samples in input dataset
according to the following condition:
}
\details{
Each resulting region r in input dataset is the contiguous intersection of at least minAcc
and at most maxAcc contributing regions in the samples of result dataset;
minAcc and maxAcc are called accumulation indexes.

Resulting regions may have new attributes, calculated by means of aggregate expressions over
the attributes of the contributing regions.
Jaccard Indexes are standard measures of similarity of the contributing regions,
added as default region attributes.
with grouping define instead, the samples are partitioned by groups,
each with distinct values of grouping metadata attributes (i.e., homonym attributes in the operand schemas)
and the cover operation is separately applied to each group,
yielding to one sample in the result for each group
}
