% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cover.R
\name{cover}
\alias{cover}
\title{GMQL Operation: COVER}
\usage{
cover(input_data, minAcc, maxAcc, groupBy = NULL, aggregates = NULL)
}
\arguments{
\item{input_data}{returned object from any GMQL function}

\item{minAcc}{minimum number of overlapping regions to be considered during execution}

\item{maxAcc}{maximum number of overlapping regions to be considered during execution}

\item{groupBy}{list of CONDITION objects every object contains the name of metadata to be used in semijoin,
or simple string concatenation c("cell_type","attribute_tag","size") without declaring condition.
In the latter form all metadata are considered having DEF condition
The CONDITION's available are:
\itemize{
\item{FULL: Fullname evaluation, two attributes match if they both end with value and,
if they have a further prefixes, the two prefix sequence are identical}
\item{DEF: Default evaluation, two attributes match if both end with value. }
\item{EXACT: Exact evaluation, only attributes exactly as value will match; no further prefixes are allowed. }
}
Every condition accepts only one string value. (e.g. DEF("cell_type") )}

\item{aggregates}{a list of element in the form key = 'function_aggregate'.
'function_aggregate' is an object of inherited class of class OPERATOR
The aggregate functions available are: MIN, MAX, SUM, BAG, AVG, COUNT,MEDIAN
Every operator accepts a string value, execet for COUNT that cannot have a value.
Argument of 'function_aggregate' must exist in schema
Two style are allowed:}
}
\value{
"url-like" string
}
\description{
it takes as input a dataset and returns another dataset (with a single sample, if no \emph{groupby} option is specified)
by “collapsing” the input samples and their regions according to certain rules specified by the input parameters.
The attributes of the output regions are only the region coordinates, and Jaccard indexes ( JaccardIntersect and JaccardResult).
Jaccard Indexes are standard measures of similarity of the contributing regions, added as default region attributes.
The JaccardIntersect index is calculated as the ratio between the lengths of the intersection
and of the union of the contributing regions; the JaccardResult index is calculated as the ratio
between the lengths of the result and of the union of the contributing regions.
If aggregate functions are specified, new attributes with aggregate values over schema region values;
Output metadata are the union of the input ones.
If \emph{groupby} clause is specified, the input samples are partitioned in groups,
each with distinct values of the grouping metadata attributes, and the COVER operation is separately
applied to each group, yielding to one sample in the result for each group.
Input samples that do not satisfy the \emph{groupby} condition are disregarded.
}
\examples{

\dontrun{

library(rscala)

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)
c = cover(input_data = r,2,3)
}
.
}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
\seealso{
\code{\link{summit}} \code{\link{flat}} \code{\link{histogram}}
}
