% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete dataset}
\usage{
deleteDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}
}
\description{
It deletes single private dataset from repository specified by name
}
\details{
If error occured a specific error will be printed
}
\examples{

\dontrun{

login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
deleteDataset(url,<dataset_name>)
}
}
\seealso{
\code{\link{downloadDataset}}
}
