% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{deleteDataset}
\alias{deleteDataset}
\title{Delete dataset}
\usage{
deleteDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to delete}
}
\value{
no object return
}
\description{
It deletes single private dataset from repository specified by name
}
\details{
If no error occured print "Deleted Dataset" otherwise a specific error will be printed
}
\examples{

\dontrun{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
deleteDataset(PolimiUrl,"job_test1_test101_20170604_180908_RESULT_DS")
}
""
}
\seealso{
\code{\link{downloadDataset}}
}
