% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Difference.R
\name{difference}
\alias{difference}
\title{GMQL Operation: DIFFERENCE}
\usage{
difference(left_input_data, right_input_data, joinBy = NULL,
  is_exact = FALSE)
}
\arguments{
\item{left_input_data}{"url-like" string taken from GMQL function}

\item{right_input_data}{"url-like" string taken from GMQL function}

\item{joinBy}{list of CONDITION objects where every object contains the name of metadata to be used in joinBy
The CONDITION's available are: EXACT, FULL, DEF
Every condition accepts only one string value. (e.g. DEF("cell_type") )}

\item{is_exact}{logical}
}
\value{
"url-like" string
}
\description{
It produces one sample in the result for each sample of the left operand,
by keeping the same metadata of the left operand sample and only those regions
(with their schema and values) of the left operand sample which do not intersect with any region
in the right operand sample.
The optional joinby clause is used to extract a subset of couples
from the cartesian product of two dataset Dleft x Dright
on which to apply the DIFFERENCE operator:
only those samples that have the same value for each attribute
are considered when performing the difference.
}
\examples{
\dontrun{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
test_path2 <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
r = read(test_path)
r2 = read(test_path2)
d = difference(r, r2, c("cell_type","age","cell_attribute","size"))
}

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
