% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Difference.R
\name{difference}
\alias{difference}
\title{GMQL Operation: DIFFERENCE}
\usage{
difference(left_input_data, right_input_data, joinBy = NULL,
  is_exact = FALSE)
}
\arguments{
\item{left_input_data}{returned object from any GMQL function}

\item{right_input_data}{returned object from any GMQL function}

\item{joinBy}{list of CONDITION objects where every object contains the name of metadata to be used in semijoin,
or simple string concatenation of name of metadata (e.g c("cell_type","attribute_tag","size") ) without declaring condition.
In the latter form all metadata are considered having DEF condition
The CONDITION's available are:
\itemize{
\item{\code{\link{FULL}}: Fullname evaluation, two attributes match if they both end with value and,
if they have a further prefixes, the two prefix sequence are identical}
\item{\code{\link{DEF}}: Default evaluation, two attributes match if both end with value. }
\item{\code{\link{EXACT}}: Exact evaluation, only attributes exactly as value will match; no further prefixes are allowed. }
}
Every condition accepts only one string value. (e.g. FULL("cell_type") )}

\item{is_exact}{logical value: TRUE means that the regions difference is executed only on regions in left_input_data
with exactly the same coordinates of at least one region present in right_input_data
if is_exact = FALSE the difference is executed on all regions in left_input_data that overlaps with at least one region in
right_input_data (even just one base)}
}
\value{
"url-like" string
}
\description{
It produces one sample in the result for each sample of the left operand,
by keeping the same metadata of the left operand sample and only those regions
(with their schema and values) of the left operand sample which do not intersect with any region
in the right operand sample.
The optional joinby clause is used to extract a subset of couples
from the cartesian product of two dataset Dleft x Dright
on which to apply the DIFFERENCE operator:
only those samples that have the same value for each attribute
are considered when performing the difference.
}
\examples{

\dontrun{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
test_path2 <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
r = read(test_path)
r2 = read(test_path2)

#### difference with condition
d = difference(r,r2 list("cell_type",EXACT("cell_attribute")) )

#### difference with DEF condition
d = difference(r,r2 list("cell_type","cell") )

#### select with condition
#### the first is FULL and the other ones are DEF
d = difference(r,r2, c(FULL("attribute_tag"),"cell_type","cell") )
}
.

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
