% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download Dataset}
\usage{
downloadDataset(url, datasetName, path = getwd())
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}

\item{path}{local path folder where store dataset
by defualt is R working directory}
}
\value{
no return value
}
\description{
It donwloads private dataset as zip file from repository to local path specified
}
\details{
If error occured a specific error will be printed
}
\examples{

#### download dataset in r working directory
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl,"test101","test")
downloadDataset(PolimiUrl,"job_test1_test101_20170604_180908_RESULT_DS",path = getwd())

}
