% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{downloadDataset}
\alias{downloadDataset}
\title{Download Dataset}
\usage{
downloadDataset(url, datasetName, path = getwd())
}
\arguments{
\item{url}{server address}

\item{datasetName}{dataset name to delete}

\item{path}{local path folder where store dataset
by defualt is R working directory}
}
\description{
It donwload private dataset from repository
}
\details{
If error occured stop and print error
}
\examples{

\dontrun{
login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
downloadDataset(url,<dataset_name>,path = <folder_path>)
}
}
