% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{downloadDataset}
\alias{downloadDataset}
\title{GMQL API web Service}
\usage{
downloadDataset(datasetName, path = getwd())
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
