% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Materialize.R
\name{execute}
\alias{execute}
\title{GMQL Function: EXECUTE}
\usage{
execute()
}
\value{
no returned value
}
\description{
execute GMQL query.
The function works only after invoking at least one materialize
}
\examples{

\dontrun{

library(rscala)

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)
s = select(input_data = r)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = s)
materialize(input_data = m, dir_out = test_path)
execute()
}

}
