% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesToGMQL.R
\name{exportGMQL.gdm}
\alias{exportGMQL.gdm}
\title{Create GMQL dataset from Granges or GrangesList}
\usage{
exportGMQL.gdm(..., dir_out)
}
\arguments{
\item{...}{set of Granges or a single GrangesList}

\item{dir_out}{folder path where create a folder and write all the sample files}
}
\description{
Create GMQL dataset from Granges or GrangesList
}
