% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extend.R
\name{extend}
\alias{extend}
\title{GMQL Operation: EXTEND}
\usage{
extend(metadata = NULL, input_data)
}
\arguments{
\item{metadata}{a list of element key = value. value is an object of class OPERATOR.
The functions aggregate available for extend function are: MIN,MAX,SUM,BAG,AVG,COUNT
every operator accept a string value}

\item{input_data}{url-like "string" pointer returned from GMQL function}
}
\description{
It generates new metadata attributes as result of aggregate functions applied to region attributes;
aggregate functions are applied sample by sample.
}
\examples{

startGMQL()
path = /.../dataset_name
r = read(path)
e = extend(somma = SUM("pvalue"),c = COUNT(), m = AVG("score"),input_data = r)

}
