% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{extend}
\alias{extend}
\title{GMQL Operation: EXTEND}
\usage{
extend(metadata = NULL, input_data)
}
\arguments{
\item{metadata}{a list of MetaAggregatesClass object}

\item{input_data}{string pointer taken from GMQL function}
}
\description{
GMQL Operation: EXTEND
}
