% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{GMQLlogin}
\alias{GMQLlogin}
\title{GMQL API web Service}
\usage{
GMQLlogin(username = NULL, password = NULL)
}
\arguments{
\item{username}{user name}

\item{password}{user password}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
