% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{GMQLregister}
\alias{GMQLregister}
\title{GMQL API web Service}
\usage{
GMQLregister(name, lastname, mail, username, password)
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
