% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{GMQLregister}
\alias{GMQLregister}
\title{GMQL API web Service}
\usage{
GMQLregister(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{server address}

\item{name}{user name}

\item{lastname}{user lastname}

\item{mail}{user mail}

\item{username}{user username}

\item{password}{user password}
}
\value{
autentication Token
}
\description{
Register to web Service Provider
}
\examples{
GMQLregister(url = http://...../...)

}
