% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gdm}
\alias{importGMQL.gdm}
\title{Create GrangesList from dataset GMQL in GDM (delimited / tabulated) format file}
\usage{
importGMQL.gdm(datasetName)
}
\arguments{
\item{datasetName}{GMQL dataset folder path}
}
\description{
It create a GrangesList from GMQL samples in dataset
It reads only sample files in GDM format
}
\examples{

\dontrun{
path = "<path_folder_input>"
grl = importGMQL.gdm(path)
}

}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gdm} }
}
