% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gtf}
\alias{importGMQL.gtf}
\title{Create GRangesList from GMQL Dataset}
\usage{
importGMQL.gtf(datasetName)
}
\arguments{
\item{datasetName}{single string GMQL dataset folder path}
}
\value{
GrangesList containing all GMQL sample In dataset
}
\description{
It create a GrangesList from GMQL samples in dataset
It reads only sample files in GTF format
}
\examples{

test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
grl = importGMQL.gtf(test_path)


}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gdm} }
}
