% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLtoGRanges.R
\name{importGMQL.gtf}
\alias{importGMQL.gtf}
\title{Create GRangesList from GMQL Dataset}
\usage{
importGMQL.gtf(datasetName)
}
\arguments{
\item{datasetName}{GMQL dataset folder path}
}
\description{
It create a GrangesList from GMQL samples in dataset
It reads only sample files in GTF format
}
\examples{

\dontrun{
path = "<path_folder_input>"
grl = importGMQL.gtf(path)
}

}
\seealso{
\code{\link{exportGMQL.gdm}} \code{\link{exportGMQL.gtf}} \code{\link{importGMQL.gdm} }
}
