% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Join.R
\name{join}
\alias{join}
\title{GMQL Operation: JOIN}
\usage{
join(genometric_predicate = NULL, joinBy = NULL, output = CONTIG(),
  right_input_data, left_input_data)
}
\arguments{
\item{genometric_predicate}{is a concatenation of distal conditions by means of logical ANDs}

\item{joinBy}{list of \code{\link{CONDITION}} objects where every object contains the name of metadata to be used in joinBy
The CONDITION's available are: EXACT, FULLNAME, DEFAULT.
Every condition accepts only one string value. (e.g. DEFAULT("cell_type") )}

\item{output}{is \code{\link{BUILDER}} object: one of four different values that declare which region is given in
output for each input pair of left dataset and right dataset regions satisfying the genometric predicate:
\itemize{
\item{LEFT}
\item{RIGHT}
\item{INTERSECTION}
\item{CONTIG}
}}

\item{right_input_data}{"url-like" string taken from GMQL function}

\item{left_input_data}{"url-like" string taken from GMQL function}
}
\description{
It takes in input two datasets, respectively known as nchor (left) and experiment (right) and returns
a dataset of samples consisting of regions extracted from the operands according to the specified condition
(a.k.a genometric_predicate).
The number of generated output samples is the Cartesian product of the number of samples
in the anchor and in the experiment dataset (if joinBy is not specified).
The output metadata are the union of the input metadata, with their attribute names prefixed with
left or right respectively.
}
\examples{
\dontrun{
initGMQL("gtf")
path = "/<path_to_your_folder>/<your_dataset_name>"
r = read(path)
c = cover(2,3,input_data = r)
j = join (list(list(UP(),MD(1)), list(DOWN(),DGE(5000))),right_input_data = r,left_input_data = c)

}


}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
