% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{login.GMQL}
\alias{login.GMQL}
\title{GMQL API web Service}
\usage{
login.GMQL(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{server address}

\item{username}{user name used on registration}

\item{password}{password used on registration}
}
\description{
Log in to GMQL web service as guest or registered user
trough username and password
}
\details{
After login you receive an authenticated token.
This token is saved in Global environment and used to perform the other REST call

\dontrun{
login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
}
}
\examples{

}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
