% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{login.GMQL}
\alias{login.GMQL}
\title{Log in to GMQL web service}
\usage{
login.GMQL(url, username = NULL, password = NULL)
}
\arguments{
\item{url}{server address}

\item{username}{single string name used on registration}

\item{password}{single string password used on registration}
}
\description{
Log in to GMQL web service as guest or registered user
trough username and password
}
\details{
if username and password are NULL you will log in as guest
After login you receive an authenticated token.
This token is saved in Global environment and used to perform the other REST call
}
\examples{

\dontrun{

url <- <http_server_address>
login.GMQL(url)
login.GMQL(url, username="pippo",password="baudo")
}

}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
