% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{logout.GMQL}
\alias{logout.GMQL}
\title{Logout from GMQL}
\usage{
logout.GMQL(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}
}
\description{
Logout from GMQL REST services suite
}
\details{
After logout the authentication token will be invalidated.
The authentication token is removed from Global environment
If error occured a specific error will be printed
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
logout.GMQL()
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
