% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{logout.GMQL}
\alias{logout.GMQL}
\title{Log out to GMQL web services}
\usage{
logout.GMQL(url)
}
\arguments{
\item{url}{server address}
}
\description{
Log out to GMQL web services
}
\details{
After logout you will not be able to call any services.
The authneticated token is removed from Global environment
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
logout.GMQL()
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
