% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{materialize}
\alias{materialize}
\title{Materialize new dataset}
\usage{
materialize(input_data, dir_out = "/Users/simone/Downloads/res")
}
\arguments{
\item{input_data}{string pointer taken from GMQL function}

\item{dir_out}{out folder path (e.g /Users/../../foldername)}
}
\description{
Materialize new dataset
}
