% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{merge}
\alias{merge}
\title{GMQL Operation: MERGE}
\usage{
merge(groupBy = NULL, input_data)
}
\arguments{
\item{input_data}{string pointer taken from GMQL function}

\item{metadata}{a list of MetaAggregate object}
}
\description{
It builds a dataset consisting of a single sample having as regions
all the regions of the input data (without altering their
coordinates, even when overlapping) and as metadata the union of all
the attribute-values of the input samples.
When a groupBy clause is present, the samples are partitioned by groups,
each with distinct values of grouping metadata attributes.
The merge operation is separately applied to each group,
yielding to one sample in the result for each group.
}
\examples{
r = read(path)
m = merge(groupBy = c(),input_data = r)
m = merge(input_data = r)

}
