% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge.R
\name{merge}
\alias{merge}
\title{GMQL Operation: MERGE}
\usage{
merge(input_data, groupBy = NULL)
}
\arguments{
\item{input_data}{"url-like" string taken from GMQL function}

\item{groupBy}{a vector of strings specifying grouping criteria}
}
\value{
"url-like" string
}
\description{
It builds a dataset consisting of a single sample having as many regions
as the numebr of regions of the input data and as many metadata as the union of
the 'attribute-value' tuples of the input samples.
A groupby clause can be specified on metadata: the samples are then partitioned in groups,
each with a distinct value of the grouping metadata attributes.
The operation is separately applied to each group, yielding one sample in the result for each group.
Samples whose names are not present in the grouping metadata parameter are disregarded.
}
\examples{

\dontrun{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)
m = merge(groupBy = c("antibody_targer","cell_karyotype"),input_data = r)
}

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
