% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{metadataFromSample}
\alias{metadataFromSample}
\title{Show metadata list from dataset sample}
\usage{
metadataFromSample(url, datasetName, sampleName)
}
\arguments{
\item{url}{server address}

\item{datasetName}{dataset name to delete}

\item{sampleName}{sample name}
}
\value{
list of metadata
}
\description{
It retrieve metadata for a specific sample in dataset
}
