% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering-class.R
\name{ORDER}
\alias{ORDER}
\title{ORDER object class}
\usage{
ORDER(value)
}
\description{
ORDER object available are:
\itemize{
\item{ASC}
\item{DESC}
}
}
\details{
you never use a parent class ORDER()
}
