% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{order}
\alias{order}
\title{GMQL operation: ORDER}
\usage{
order(metadata_order = NULL, mtop = 0, mtopg = 0, regions_order = NULL,
  rtop = 0, rtopg = 0, input_data)
}
\arguments{
\item{metadata_order}{list of ORDER objects where every object contains the name of metadata
The ORDER's available are: ASC, DESC.
Every condition accepts only one string value. (e.g. ASC("cell_type") )}

\item{mtop}{integer value specifying the first k samples.
default is 0 that means every sample must be considered}

\item{mtopg}{integer value specifying the first j samples in each group.
default is 0 that means every sample must be considered}

\item{regions_order}{list of ORDER objects where every object contains the name of region schema value
The ORDER's available are: ASC, DESC.
Every condition accepts only one string value. (e.g. DESC("pvalue") )}

\item{rtop}{integer value specifying the first m samples in each group.
default is 0 that means every sample must be considered}

\item{rtopg}{integer value specifying the first i samples in each group.
default is 0 that means every sample must be considered}

\item{input_data}{"url-like" string taken from GMQL function}
}
\description{
It is used to order either samples or sample regions or both,
according to a set of metadata and/or region attributes, and/or region coordinates.
Order can be specified as ascending / descending for every attribute
The number of samples and their regions remain the same (unless mtop/rtop parameters specified)
but a new ordering metadata and/or region attribute is added.
Sorted samples or regions have a new attribute "order", added to either metadata, or regions,
or both of them as specified in input
The input mtop = k and rtop = m extracts the first k samples and m regions respectively,
the clause mtopg = k and rtopg = m performs grouping operation,
grouping by identical values of ordering attributes
and then selects the first k samples or regions of each group
}
\details{
mtop, mtopg, rtop and rtopg are normally numbers: if you specify a vector,
only the first element will be used
}
\examples{

startGMQL()
path = /.../dataset_name
r = read(path)
c = cover(2,3,input_data = r)
s = select("NOT(Patient_age < 70 AND provider=='Polimi')",input_dat = r)
s = select("NOT(Patient_age < 70)",region_predicate = "NOT(qValue > 0.001)",
semi_join = list(EXACT("cell_type"),EXACT("age")),semi_join_dataset = c,input_data = r )

o = order(DESC(Region_Count), mtop = 2, input_data = s)
o = order(list(DESC(Region_Count)),regions_order = list(DESC(MutationCount),ASC(pvalue)),
mtop = 5,rtopg = 1, input_data = c)


}
