% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{order}
\alias{order}
\title{GMQL operation: ORDER}
\usage{
order(input_data, metadata_ordering = NULL, mtop = 0, mtopg = 0,
  regions_ordering = NULL, rtop = 0, rtopg = 0)
}
\arguments{
\item{input_data}{"url-like" string taken from GMQL function}

\item{metadata_ordering}{list of \code{\link{ORDER}} objects where every object contains the name of metadata
The ORDER's available are: ASC, DESC
Every condition accepts only one string value. (e.g. ASC("cell_type") )}

\item{mtop}{integer value specifying the first k samples.
default is 0 that means every sample must be considered}

\item{mtopg}{integer value specifying the first j samples in each group.
default is 0 that means every sample must be considered}

\item{regions_ordering}{list of \code{\link{ORDER}} objects where every object contains the name of region schema value
The ORDER's available are: ASC, DESC.
Every condition accepts only one string value. (e.g. DESC("pvalue") )}

\item{rtop}{integer value specifying the first m samples in each group.
default is 0 that means every sample must be considered}

\item{rtopg}{integer value specifying the first i samples in each group.
default is 0 that means every sample must be considered}
}
\value{
"url-like" string
}
\description{
It is used to order either samples or sample regions or both,
according to a set of metadata and/or region attributes, and/or region coordinates.
Order can be specified as ascending / descending for every attribute
The number of samples and their regions remain the same (unless mtop/rtop parameters specified)
but a new ordering metadata and/or region attribute is added.
Sorted samples or regions have a new attribute "order", added to either metadata, or regions,
or both of them as specified in input
The input mtop = k and rtop = m extracts the first k samples and m regions respectively,
the clause mtopg = k and rtopg = m performs grouping operation,
grouping by identical values of ordering attributes
and then selects the first k samples or regions of each group
}
\details{
mtop, mtopg, rtop and rtopg are normally numbers: if you specify a vector,
only the first element will be used
mtop and mtopg are mutalbe exclusive, so rtop and rtopg
if you specified both, only mtop (rtop) is taken
}
\examples{
\dontrun{

initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)
s = select()
o = order(DESC(Region_Count), mtop = 2, input_data = s)
}

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
