% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{project}
\alias{project}
\title{GMQL Operation: PROJECT}
\usage{
project()
}
\arguments{
\item{predicate}{string made up by logical oepration: AND,OR,NOT}

\item{region}{region}

\item{semijoin}{semijoin}

\item{input_data}{string pointer taken from GMQL function}
}
\description{
GMQL Operation: PROJECT
}
