% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{project}
\alias{project}
\title{GMQL Operation: PROJECT}
\usage{
project(metadata = NULL, regions = NULL, input_data)
}
\arguments{
\item{input_data}{string pointer taken from GMQL function}

\item{predicate}{string made up by logical oepration: AND,OR,NOT}

\item{region}{region}

\item{semijoin}{semijoin}
}
\description{
GMQL Operation: PROJECT
}
