% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(DatasetPathFolder)
}
\arguments{
\item{DatasetPathFolder}{folder path for GMQL dataset}
}
\value{
url-like string "pointer" to dataset
}
\description{
Read a GMQL dataset from disk
}
\examples{
startGMQL()
path = /.../dataset_name
r = read(path)


}
