% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(DatasetPathFolder, parser = CustomParser())
}
\arguments{
\item{DatasetPathFolder}{folder path for GMQL dataset}

\item{parser}{\code{\link{PARSER}} objects used to parsing dataset files
The PARSER's available are: BedParser, ANNParser, BroadProjParser, BasicParser,
NarrowPeakParser, RnaSeqParser, CustomParser.
Default is CustomParser.}
}
\value{
"url-like" string to dataset
}
\description{
Read a GMQL dataset or any other folder containig some homogenus sample
from disk, saving in Scala memory that can be referenced in R
}
\details{
Normally a GMQL dataset contains an XML schema file that contains
name of column header. (e.g chr, start, stop, strand)
The CustomParser read this XML schema; if you already know what kind of schema your files are use one of
the parser defined without reading any XML schema
}
\examples{

\dontrun{
startGMQL()
path = "/<path_to_your_folder>/<your_dataset_name>"
r = read(path)
r = read(path,BedParser())
r = read(path,RnaSeqParser())
}

}
