% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(samples)
}
\arguments{
\item{samples}{GrangesList}
}
\value{
"url-like" string to dataset
}
\description{
Read a GrangesList saving in Scala memory that can be referenced in R
}
\examples{

\dontrun{
}
""

}
