% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{read}
\alias{read}
\title{GMQL Function: READ}
\usage{
read(DatasetFolder, parser = "CustomParser", is_local = TRUE)
}
\arguments{
\item{DatasetFolder}{folder path for GMQL dataset or datasetname on repository}

\item{parser}{string used to parsing dataset files
The Parser's available are:
\itemize{
\item{BedParser}
\item{ANNParser}
\item{BroadProjParser}
\item{BedParser}
\item{NarrowPeakParser}
\item{RnaSeqParser}
\item{CustomParser.}
}
Default is CustomParser.}

\item{is_local}{logical value indicates if dataset is a local or remote dataset
if the remote processing is off you cannot use a remote repository (error occured)}
}
\value{
"url-like" string to dataset
}
\description{
Read a GMQL dataset or any other folder containig some homogenus sample
from disk, saving in Scala memory that can be referenced in R
}
\details{
Normally a GMQL dataset contains an XML schema file that contains
name of column header. (e.g chr, start, stop, strand)
The CustomParser read this XML schema; if you already know what kind of schema your files are use one of
the parser defined without reading any XML schema
}
\examples{

\dontrun{

### with CustomParser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path)

 #### with an other Parser
initGMQL("gtf")
test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
r = read(test_path,"ANNParser")
}

}
