% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQLOperation.R
\name{read}
\alias{read}
\title{Read Dataset from Disk}
\usage{
read(DatasetPathFolder)
}
\arguments{
\item{DatasetPathFolder}{folder path (e.g /Users/../../foldername)}
}
\value{
string "pointer" to dataset
}
\description{
Allow to read a GMQL Dataset from disk
}
