% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{regionFromSample}
\alias{regionFromSample}
\title{Shows regions from a dataset sample}
\usage{
regionFromSample(url, datasetName, sampleName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}

\item{sampleName}{sample name we want to get}
}
\value{
Granges data containing regions of sample
}
\description{
It retrieves regions for a specific sample (whose name is specified in the paramter "sampleName")
in a specific dataset (whose name is specified in the paramter "datasetName")
}
\details{
If error occured a specific error will be printed

\dontrun{

login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
regionFromSample(url,"<dataset_name>",<dataset_sample_name>)
}
}
