% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{regionFromSample}
\alias{regionFromSample}
\title{Show regions from dataset sample}
\usage{
regionFromSample(url, datasetName, sampleName)
}
\arguments{
\item{url}{server address}

\item{datasetName}{dataset name to delete}

\item{sampleName}{sample name}
}
\value{
Granges containing regions
}
\description{
It retrieve regions for a specific sample in dataset
}
