% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{regionFromSample}
\alias{regionFromSample}
\title{GMQL API web Service}
\usage{
regionFromSample()
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
