% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{register.GMQL}
\alias{register.GMQL}
\title{GMQL API web Service}
\usage{
register.GMQL(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{server address}

\item{name}{user name}

\item{lastname}{user lastname}

\item{mail}{user mail}

\item{username}{user username}

\item{password}{user password, unfortunatley the password is not encrypted and not hide}
}
\description{
Register to Web Service Provider
}
\examples{
\dontrun{
url <- <http_server_address>)
register.GMQL(url = url,"jonh","Doe","jonh@doe.com","JonhDoe46","password")
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
