% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{register.GMQL}
\alias{register.GMQL}
\title{GMQL API web Service}
\usage{
register.GMQL(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{server address}

\item{name}{user name}

\item{lastname}{user lastname}

\item{mail}{user mail}

\item{username}{user username}

\item{password}{user password}
}
\description{
Register to web Service Provider
}
\examples{
\dontrun{
register.GMQL(url = http://...../...)
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
