% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authOperations.R
\name{register.GMQL}
\alias{register.GMQL}
\title{Register to GMQL}
\usage{
register.GMQL(url, name, lastname, mail, username, password)
}
\arguments{
\item{url}{server address}

\item{name}{first name}

\item{lastname}{lastname}

\item{mail}{email}

\item{username}{username}

\item{password}{password}
}
\description{
Register to GMQL rest services suite
}
\details{
After registration you receive an authenticated token.
This token is saved in Global environment and used to perform the other REST call
}
\examples{

\dontrun{

url <- <http_server_address>)
register.GMQL(url = url,"jonh","Doe","jonh@doe.com","JonhDoe46","password")
}
}
\seealso{
\code{\link{register.GMQL}} \code{\link{logout.GMQL}}
}
