% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{runQuery.fromfile}
\alias{runQuery.fromfile}
\title{Run GMQL query}
\usage{
runQuery.fromfile(url, fileName, filePath, output_gtf = TRUE)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{fileName}{single string name of the file}

\item{filePath}{single string path of txt files where you wrote a query}

\item{output_gtf}{logical: file format for sample generating from query
the possiblities are: GTF or TAB}
}
\value{
no object return
}
\description{
It runs a GMQL query from file
}
\details{
If error occured a specific error will be printed
}
\examples{

## run query: output GTF

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
test_path <- system.file("example",package = "GMQL")
test_query <- file.path(test_path, "query1.txt")
runQuery(PolimiUrl, "query_1", test_query, output_gtf = FALSE)

## run query: output GDM (tabulated)

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl)
test_path <- system.file("example",package = "GMQL")
test_query <- file.path(test_path, "query1.txt")
runQuery(PolimiUrl, "query_1", test_query, output_gtf = TRUE)

}
