% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{runQuery}
\alias{runQuery}
\title{GMQL API web Service}
\usage{
runQuery(url, fileName, query, output_gtf = T)
}
\arguments{
\item{url}{server address}

\item{fileName}{name of file}

\item{query}{text of query}

\item{output_gtf}{logical: file format for sample generating from query
the possiblities are: GTF or TAB}
}
\description{
It runs a query
}
