% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{saveQuery.fromfile}
\alias{saveQuery.fromfile}
\title{GMQL API web Service}
\usage{
saveQuery.fromfile(url, queryName, filePath)
}
\arguments{
\item{url}{server address}

\item{queryName}{name of query}

\item{filePath}{local file path where you write GMQL query}
}
\description{
It save the GMQL queries to repository
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
saveQuery.fromfile(url,"prova1","<query_file_path>")
logout.GMQL()
}

}
\seealso{
\code{\link{saveQuery}}
}
