% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowse.R
\name{saveQuery}
\alias{saveQuery}
\title{GMQL API web Service}
\usage{
saveQuery(nameQuery, queryTxt)
}
\arguments{
\item{nameQuery}{name of query}

\item{queryTxt}{text of query}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
