% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{saveQuery}
\alias{saveQuery}
\title{GMQL API web Service}
\usage{
saveQuery(url, queryName, queryTxt)
}
\arguments{
\item{url}{server address}

\item{queryName}{name of query}

\item{queryTxt}{text of GMQL query}
}
\description{
It save the GMQL queries to repository
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
saveQuery(url,"prova1","<query_text>")
logout.GMQL()
}

}
\seealso{
\code{\link{showQueries}} \code{\link{saveQuery.fromfile}}
}
