% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{saveQuery}
\alias{saveQuery}
\title{Save GMQL query}
\usage{
saveQuery(url, queryName, queryTxt)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{queryName}{single string name of query}

\item{queryTxt}{single string text of GMQL query}
}
\description{
It saves the GMQL query into repository
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>
login.GMQL(url)
saveQuery(url,"prova1","<query_text>")
logout.GMQL()
}

}
\seealso{
\code{\link{showQueries}} \code{\link{saveQuery.fromfile}}
}
