% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{saveQuery}
\alias{saveQuery}
\title{GMQL API web Service}
\usage{
saveQuery(url = "http://genomic.elet.polimi.it/gmql-rest/query/", queryName,
  queryTxt)
}
\arguments{
\item{url}{server address}

\item{queryName}{query name}

\item{queryTxt}{query text}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
