% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{select}
\alias{select}
\title{GMQL Operation: SELECT}
\usage{
select(input_data, predicate = NULL, region_predicate = NULL,
  semi_join = NULL, semi_join_dataset = NULL)
}
\arguments{
\item{input_data}{"url-like" string taken from GMQL function}

\item{predicate}{string predicate made up by logical oepration: AND,OR,NOT on metadata attribute}

\item{region_predicate}{string predicate made up by logical operation: AND,OR,NOT on schema region values}

\item{semi_join}{list of \code{\link{CONDITION}} objects every object contains the name of metadata to be used in semijoin,
The CONDITION's available are: FULL, DEF, EXACT
Every condition accepts only one string value. (e.g. DEF("cell_type") )}

\item{semi_join_dataset}{"url-like" string taken from GMQL function used in semijoin}
}
\description{
It extracts a subset of samples from the input dataset.
It returns all the samples satisfying the predicate on metadata.
If regions are specified, returns regions satisfying the predicate on regions.
If semijoin clauses are specified they are applied, too.
When semijoin is defined, it extracts those samples containing all metadata attribute defined in semijoin clause
with at least one metadata value in common with semi join dataset
If no metadata in common beetween input dataset and semi join dataset, no sample is extracted
}
\examples{
\dontrun{

initGMQL("tab")
path = "/<path_to_your_folder>/<your_dataset_name>"
r = read(path)
c = cover(2,3,input_data = r)
s = select("NOT(Patient_age < 70 AND provider=='Polimi')",input_dat = r)
s = select(region_predicate = "NOT(qValue > 0.001)", semi_join = list("cell_type","age",EXACT("cell")),
semi_join_dataset = c,input_data = r )
s = select("NOT(Patient_age < 70)", semi_join = c("cell_type","age","cell_attribute","size"),
semi_join_dataset = c,input_data = r )
}

}
\references{
\url{http://www.bioinformatics.deib.polimi.it/genomic_computing/GMQL/doc/GMQLUserTutorial.pdf}
}
