% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showDatasets}
\alias{showDatasets}
\title{GMQL API web Service}
\usage{
showDatasets(url = "http://genomic.elet.polimi.it/gmql-rest/datasets")
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
