% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{showDatasets}
\alias{showDatasets}
\title{Show Dataset}
\usage{
showDatasets(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}
}
\value{
list of datasets.

Every dataset in the list is identified by:
\itemize{
\item{name: name of dataset}
\item{owner: pulbic or name of the user}
}
}
\description{
It show all GMQL dataset stored in repository
}
\details{
If error occured a specific error will be printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl)
list <- showDatasets(PolimiUrl)

}
\seealso{
\code{\link{deleteDataset}}
}
