% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showDatasets}
\alias{showDatasets}
\title{Show Dataset}
\usage{
showDatasets(url)
}
\arguments{
\item{url}{server address}
}
\value{
list of datasets
every dataset in the list is identified by:
\itemize{
\item{name}
\item{owner}
}
}
\description{
It show all GMQL dataset stored in repository
}
\details{
If error occured stop and print error
}
\examples{

\dontrun{

url <- <http_server_address>
login.GMQL(url = <http_server_address>)
list <- showDatasets(url)
}
}
\seealso{
\code{\link{deleteDataset}}
}
