% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{showQueries}
\alias{showQueries}
\title{GMQL API web Service}
\usage{
showQueries(url)
}
\arguments{
\item{url}{server address}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
