% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{showQueries}
\alias{showQueries}
\title{Shows all Queries}
\usage{
showQueries(url)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}
}
\value{
list of queries
Every query in the list is identified by:
\itemize{
\item{name: name of query}
\item{text: text of GMQL query}
}
}
\description{
It shows all the GMQL query saved to repository
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
list <- showQueries(url)
logout.GMQL()
}

}
\seealso{
\code{\link{saveQuery}}
}
