% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBrowser.R
\name{showQueries}
\alias{showQueries}
\title{GMQL API web Service}
\usage{
showQueries(url)
}
\arguments{
\item{url}{server address}
}
\value{
list of queries
}
\description{
It show all the query saved to repository
}
\details{
if error occured print the content error
}
\examples{

\dontrun{

url <- <http_server_address>)
login.GMQL(url, username="pippo",password="baudo")
list <- showQueries(url)
logout.GMQL()
}

}
\seealso{
\code{\link{saveQuery}}
}
