% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showSamplesFromDataset}
\alias{showSamplesFromDataset}
\title{GMQL API web Service}
\usage{

  showSamplesFromDataset(url = "http://genomic.elet.polimi.it/gmql-rest/datasets/",
  datasetName)
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
