% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{showSchemaFromDataset}
\alias{showSchemaFromDataset}
\title{Show dataset schema}
\usage{
showSchemaFromDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}
}
\value{
list of region schema fields.
Every field in the list is identified by:
\itemize{
\item{name: name of field (e.g. chr, start, end, strand ...)}
\item{fieldType: (e.g STRING, DOUBLE ...)}
}
}
\description{
It shows the region schema of a specific GMQL dataset
}
\details{
If error occured a specific error will be printed
}
\examples{

\dontrun{
login.GMQL(url = <http_server_address>)
url <- <http_server_address>)
list <- showSchemaFromDataset(url,<dataset_name>)
}
}
