% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{showSchemaFromDataset}
\alias{showSchemaFromDataset}
\title{Show dataset schema}
\usage{
showSchemaFromDataset(url, datasetName)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get
if the dataset is a public dataset, we have to add "public." as prefix, as shown in the example below}
}
\value{
list of region schema fields.

Every field in the list is identified by:
\itemize{
\item{name: name of field (e.g. chr, start, end, strand ...)}
\item{fieldType: (e.g STRING, DOUBLE ...)}
}
}
\description{
It shows the region schema of a specific GMQL dataset
}
\details{
If error occured a specific error will be printed
}
\examples{

### show schema of public dataset
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl)
list <- showSchemaFromDataset(PolimiUrl,"public.GRCh38_ENCODE_BROAD_MAY_2017")

}
