% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOp.R
\name{stopJob}
\alias{stopJob}
\title{Stop a job}
\usage{
stopJob(url, job_id)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{job_id}{single string id of the job}
}
\value{
no object return
}
\description{
It stops a specific current job
}
\details{
If error occured a specific error will be printed
}
\examples{

PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(PolimiUrl,"test101","test")
list_jobs <- showJobs(PolimiUrl)
jobs_1 <- list_jobs$jobs[[1]]
stopJob(PolimiUrl,jobs_1)

}
