% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMQL4TFarm.R
\name{TFARMatrix}
\alias{TFARMatrix}
\title{GMQL to TFARM matrix}
\usage{
TFARMatrix(GMQL_dataset_path, metadata = NULL, metadata_prefix = NULL,
  regions = NULL)
}
\arguments{
\item{GMQL_dataset_path}{aaa}

\item{metadata}{aaaaaa   aaa}

\item{metadata_prefix}{a a a a a a a}

\item{regions}{a a a a a a}
}
\value{
TFARM matrix
}
\description{
create matrix for TFARM package
}
\examples{


test_path <- system.file("example","DATA_SET_VAR_GTF",package = "GMQL")
TFARMatrix(test_path,regions = c("pvalue","peak"))


}
