% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmql4tfarm.R
\name{TFARMatrix}
\alias{TFARMatrix}
\title{GMQL to TFARM matrix
solo sample con ugual numero di righe}
\usage{
TFARMatrix(GMQL_dataset_path, metadata = NULL, metadata_prefix = NULL,
  regions = NULL)
}
\arguments{
\item{regions}{}
}
\description{
GMQL to TFARM matrix
solo sample con ugual numero di righe
}
\details{

}
\examples{

\dontrun{
}

}
