% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOperation.R
\name{traceJob}
\alias{traceJob}
\title{GMQL API web Service}
\usage{
traceJob(url, job_id)
}
\arguments{
\item{username}{user name}

\item{password}{user password}
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
