% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{uploadSamples}
\alias{uploadSamples}
\title{Upload dataset}
\usage{
uploadSamples(url, datasetName, folderPath, schemaName = NULL, isGMQL = T)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}

\item{folderPath}{local path to the folder containing the samples}

\item{schemaName}{name of schema used to parse the samples
schemaName available are:
\itemize{
\item{NARROWPEAK}
\item{BROADPEAK}
\item{VCF}
\item{BED}
\item{BEDGRAPH}
}
if schema is NULL it looking for a XML schema file to read}

\item{isGMQL}{logical value indicating whether is GMQL dataset or not}
}
\description{
It uploads folder containing dataset files: a new dataset will be created on repository
}
\details{
If error occured a specific error will be printed
}
\examples{

\dontrun{
url <- <http_server_address>)
login.GMQL(url = <http_server_address>)
uploadSamples(url,"prova1",folderPath = <folder_path>)
}

}
