% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{uploadSamples}
\alias{uploadSamples}
\title{GMQL API web Service}
\usage{
uploadSamples(url, datasetName, schemaName = NULL, folderPath,
  fileExt = "gtf")
}
\details{
schema_name <- toupper(schemaName)
 switch(schema_name,
  "BED",
  "NARROWPEAK",
  "BEDGRAPH",
  "VCF",
  "BROADPEAK")
}
