% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetOp.R
\name{uploadSamples}
\alias{uploadSamples}
\title{Upload dataset}
\usage{
uploadSamples(url, datasetName, folderPath, schemaName = NULL,
  isGMQL = TRUE)
}
\arguments{
\item{url}{single string url of server: it must contain the server address and base url;
service name will be added automatically}

\item{datasetName}{name of dataset we want to get}

\item{folderPath}{local path to the folder containing the samples}

\item{schemaName}{name of schema used to parse the samples
schemaName available are:
\itemize{
\item{NARROWPEAK}
\item{BROADPEAK}
\item{VCF}
\item{BED}
\item{BEDGRAPH}
}
if schema is NULL it's looking for a XML schema file to read}

\item{isGMQL}{logical value indicating whether is GMQL dataset or not}
}
\value{
no object return
}
\description{
It uploads an folder (GMQL or not) containing dataset files:
a new dataset will be created on repository
}
\details{
If error occured a specific error will be printed
}
\examples{

## upload of GMQL dataset with no schema selection
test_path <- system.file("example","DATA_SET_VAR_GDM",package = "GMQL")
PolimiUrl = "http://genomic.elet.polimi.it/gmql-rest"
login.GMQL(url = PolimiUrl)
uploadSamples(PolimiUrl,"dataset1",folderPath = test_path)

}
