% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSetOperation.R
\name{uploadSamples}
\alias{uploadSamples}
\title{GMQL API web Service}
\usage{
uploadSamples(url = "http://genomic.elet.polimi.it/gmql-rest/datasets",
  datasetName, schemaName = NULL, ...)
}
\description{
Allow access to web service GMQL as guest or registered user
with username and password
}
